split StudentDataCheck page into multiple files
