+++ clients/team_allocation_component/src/team_allocation/interfaces/studentCheck.ts
@@ -1,0 +1,12 @@
+import { TeaseStudent } from '../interfaces/tease/student'
+
+export type StudentCheck = {
+  label: string
+  extractor: (student: TeaseStudent) => any
+  isEmpty: (value: any) => boolean
+  missingMessage: string
+  userHint: string
+  category: 'previous' | 'devices' | 'comments' | 'score' | 'language' | 'survey'
+  highLevelCategory: 'previous' | 'survey'
+  icon: React.ReactNode
+}
