+++ clients/team_allocation_component/src/team_allocation/interfaces/validationResult.ts
@@ -1,0 +1,9 @@
+export type ValidationResult = {
+  label: string
+  isValid: boolean
+  details?: string
+  category: 'previous' | 'devices' | 'comments' | 'score' | 'language' | 'survey'
+  highLevelCategory: 'previous' | 'survey'
+  completionRate: number
+  icon: React.ReactNode
+}
