+++ clients/team_allocation_component/src/team_allocation/pages/TeaseConfig/components/StudentDataCheck.tsx
@@ -1,36 +1,17 @@
-'use client'
-
 import { useEffect, useState } from 'react'
 import { useParams } from 'react-router-dom'
 import { useQuery } from '@tanstack/react-query'
-import {
-  CheckCircle,
-  AlertCircle,
-  Loader2,
-  Users,
-  ClipboardCheck,
-  RefreshCcw,
-  FileText,
-  Laptop,
-  MessageSquare,
-  Star,
-  Languages,
-  BarChart3,
-} from 'lucide-react'
-import type React from 'react'
+import { AlertCircle, Loader2, Users, RefreshCcw } from 'lucide-react'
 
 import { getAllTeaseStudents } from '../../../network/queries/getAllTeaseStudents'
 import type { TeaseStudent } from '../../../interfaces/tease/student'
+import type { ValidationResult } from '../../../interfaces/validationResult'
+import DataCompletionSummary from './DataCompletionSummary'
+import { CheckItem } from './CheckItem'
+import { checksConfig } from './ChecksConfig'
+
 import { Skeleton } from '@/components/ui/skeleton'
-import {
-  Card,
-  CardContent,
-  CardHeader,
-  CardTitle,
-  CardDescription,
-  CardFooter,
-} from '@/components/ui/card'
-import { Progress } from '@/components/ui/progress'
+import { Card, CardContent, CardFooter } from '@/components/ui/card'
 import { Badge } from '@/components/ui/badge'
 import { Button } from '@/components/ui/button'
 import {
@@ -40,304 +21,6 @@ import {
   AccordionTrigger,
 } from '@/components/ui/accordion'
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs'
-
-type ValidationResult = {
-  label: string
-  isValid: boolean
-  details?: string
-  category: 'previous' | 'devices' | 'comments' | 'score' | 'language' | 'survey'
-  highLevelCategory: 'previous' | 'survey'
-  completionRate: number
-  icon: React.ReactNode
-}
-
-type StudentCheck = {
-  label: string
-  extractor: (student: TeaseStudent) => any
-  isEmpty: (value: any) => boolean
-  missingMessage: string
-  userHint: string
-  category: 'previous' | 'devices' | 'comments' | 'score' | 'language' | 'survey'
-  highLevelCategory: 'previous' | 'survey'
-  icon: React.ReactNode
-}
-
-const checksConfig: StudentCheck[] = [
-  // Data from previous phases
-  {
-    label: 'First Name',
-    extractor: (s) => s.firstName,
-    isEmpty: (v) => !v,
-    missingMessage: 'first names',
-    userHint: 'This data should be forwarded from previous phases.',
-    category: 'previous',
-    highLevelCategory: 'previous',
-    icon: <FileText className='h-4 w-4' />,
-  },
-  {
-    label: 'Last Name',
-    extractor: (s) => s.lastName,
-    isEmpty: (v) => !v,
-    missingMessage: 'last names',
-    userHint: 'This data should be forwarded from previous phases.',
-    category: 'previous',
-    highLevelCategory: 'previous',
-    icon: <FileText className='h-4 w-4' />,
-  },
-  {
-    label: 'Gender',
-    extractor: (s) => s.gender,
-    isEmpty: (v) => !v,
-    missingMessage: 'gender info',
-    userHint: 'This data should be forwarded from previous phases.',
-    category: 'previous',
-    highLevelCategory: 'previous',
-    icon: <FileText className='h-4 w-4' />,
-  },
-  {
-    label: 'Nationality',
-    extractor: (s) => s.nationality,
-    isEmpty: (v) => !v,
-    missingMessage: 'nationality info',
-    userHint: 'This data should be forwarded from previous phases.',
-    category: 'previous',
-    highLevelCategory: 'previous',
-    icon: <FileText className='h-4 w-4' />,
-  },
-  {
-    label: 'Study Degree',
-    extractor: (s) => s.studyDegree,
-    isEmpty: (v) => !v,
-    missingMessage: 'study degree info',
-    userHint: 'This data should be forwarded from previous phases.',
-    category: 'previous',
-    highLevelCategory: 'previous',
-    icon: <FileText className='h-4 w-4' />,
-  },
-  {
-    label: 'Study Program',
-    extractor: (s) => s.studyProgram,
-    isEmpty: (v) => !v,
-    missingMessage: 'study program info',
-    userHint: 'This data should be forwarded from previous phases.',
-    category: 'previous',
-    highLevelCategory: 'previous',
-    icon: <FileText className='h-4 w-4' />,
-  },
-  {
-    label: 'Semester',
-    extractor: (s) => s.semester,
-    isEmpty: (v) => v === 0,
-    missingMessage: 'semester info',
-    userHint: 'This data should be forwarded from previous phases.',
-    category: 'previous',
-    highLevelCategory: 'previous',
-    icon: <FileText className='h-4 w-4' />,
-  },
-
-  // Devices
-  {
-    label: 'Devices',
-    extractor: (s) => s.devices,
-    isEmpty: (arr) => arr.length === 0,
-    missingMessage: 'devices information',
-    userHint:
-      'Include a question about what kind of devices (laptop, tablet) students can use during the course.',
-    category: 'devices',
-    highLevelCategory: 'previous',
-    icon: <Laptop className='h-4 w-4' />,
-  },
-
-  // Comments
-  {
-    label: 'Tutor Comments',
-    extractor: (s) => s.tutorComments,
-    isEmpty: (arr) => !arr || arr.length === 0,
-    missingMessage: 'tutor comments',
-    userHint: 'Ensure tutors can provide comments on student performance.',
-    category: 'comments',
-    highLevelCategory: 'previous',
-    icon: <MessageSquare className='h-4 w-4' />,
-  },
-  {
-    label: 'Student Comments',
-    extractor: (s) => s.studentComments,
-    isEmpty: (arr) => !arr || arr.length === 0,
-    missingMessage: 'student comments',
-    userHint: 'Allow students to provide feedback or comments.',
-    category: 'comments',
-    highLevelCategory: 'previous',
-    icon: <MessageSquare className='h-4 w-4' />,
-  },
-
-  // ScoreLevel
-  {
-    label: 'Score Level',
-    extractor: (s) => s.introCourseProficiency,
-    isEmpty: (v) => v === undefined || v === null,
-    missingMessage: 'score level information',
-    userHint: 'Ensure score levels are assigned to students.',
-    category: 'score',
-    highLevelCategory: 'previous',
-    icon: <Star className='h-4 w-4' />,
-  },
-
-  // Language Proficiency
-  {
-    label: 'English Proficiency',
-    extractor: (s) => s.languages?.find((l) => l.language === 'English')?.proficiency,
-    isEmpty: (v) => !v,
-    missingMessage: 'English proficiency levels',
-    userHint:
-      'Add application question with possible answers "A1/A2, B1/B2, C1/C2, Native" and export with access key "language_proficiency_english".',
-    category: 'language',
-    highLevelCategory: 'previous',
-    icon: <Languages className='h-4 w-4' />,
-  },
-  {
-    label: 'German Proficiency',
-    extractor: (s) => s.languages?.find((l) => l.language === 'German')?.proficiency,
-    isEmpty: (v) => !v,
-    missingMessage: 'German proficiency levels',
-    userHint:
-      'Add application question with possible answers "A1/A2, B1/B2, C1/C2, Native" and export with access key "language_proficiency_german".',
-    category: 'language',
-    highLevelCategory: 'previous',
-    icon: <Languages className='h-4 w-4' />,
-  },
-
-  // Survey Results
-  {
-    label: 'Survey Submission',
-    extractor: (s) => s.projectPreferences,
-    isEmpty: (arr) => arr.length === 0,
-    missingMessage: 'survey submissions',
-    userHint: 'Track whether students have submitted the required survey.',
-    category: 'survey',
-    highLevelCategory: 'survey',
-    icon: <BarChart3 className='h-4 w-4' />,
-  },
-]
-
-const CheckItem = ({ check }: { check: ValidationResult }) => {
-  return (
-    <Card className={`border-l-4 ${check.isValid ? 'border-l-green-500' : 'border-l-yellow-500'}`}>
-      <CardContent className='flex items-center gap-4 p-4'>
-        <div className='flex-shrink-0'>
-          {check.isValid ? (
-            <CheckCircle className='h-6 w-6 text-green-500' />
-          ) : (
-            <AlertCircle className='h-6 w-6 text-yellow-500' />
-          )}
-        </div>
-        <div className='flex-grow'>
-          <div className='flex items-center justify-between'>
-            <div className='flex items-center gap-2'>
-              {check.icon}
-              <h3 className='font-medium'>{check.label}</h3>
-            </div>
-            <Badge
-              variant={check.isValid ? 'default' : 'outline'}
-              className={
-                check.isValid ? 'bg-green-100 text-green-800' : 'bg-yellow-100 text-yellow-800'
-              }
-            >
-              {check.isValid ? 'Complete' : `${check.completionRate}% Complete`}
-            </Badge>
-          </div>
-          <p className='mt-1 text-sm text-muted-foreground'>
-            {check.details || 'All students have provided this information.'}
-          </p>
-        </div>
-      </CardContent>
-    </Card>
-  )
-}
-
-const DataCompletionSummary = ({ checks }: { checks: ValidationResult[] }) => {
-  const totalChecks = checks.length
-  const completedChecks = checks.filter((check) => check.isValid).length
-  const completionPercentage = Math.round((completedChecks / totalChecks) * 100)
-
-  const highLevelCategoryStats = {
-    previous: {
-      total: checks.filter((c) => c.highLevelCategory === 'previous').length,
-      completed: checks.filter((c) => c.highLevelCategory === 'previous' && c.isValid).length,
-    },
-    survey: {
-      total: checks.filter((c) => c.highLevelCategory === 'survey').length,
-      completed: checks.filter((c) => c.highLevelCategory === 'survey' && c.isValid).length,
-    },
-  }
-
-  return (
-    <Card className='mb-6'>
-      <CardHeader>
-        <CardTitle className='flex items-center gap-2'>
-          <ClipboardCheck className='h-5 w-5' />
-          Data Completion Summary
-        </CardTitle>
-        <CardDescription>Overall student data completeness and category breakdown</CardDescription>
-      </CardHeader>
-      <CardContent>
-        <div className='mb-6'>
-          <div className='flex justify-between mb-2'>
-            <span className='text-sm font-medium'>Overall Completion</span>
-            <span className='text-sm font-medium'>{completionPercentage}%</span>
-          </div>
-          <Progress value={completionPercentage} className='h-2' />
-        </div>
-
-        <div className='grid grid-cols-1 md:grid-cols-2 gap-4'>
-          <Card className='border-amber-100'>
-            <CardHeader className='pb-2'>
-              <CardTitle className='text-sm font-medium text-amber-700 flex items-center gap-2'>
-                <FileText className='h-4 w-4' />
-                Data from Previous Phases
-              </CardTitle>
-            </CardHeader>
-            <CardContent>
-              <div className='text-2xl font-bold'>
-                {Math.round(
-                  (highLevelCategoryStats.previous.completed /
-                    highLevelCategoryStats.previous.total) *
-                    100,
-                )}
-                %
-              </div>
-              <p className='text-xs text-muted-foreground'>
-                {highLevelCategoryStats.previous.completed} of{' '}
-                {highLevelCategoryStats.previous.total} checks complete
-              </p>
-            </CardContent>
-          </Card>
-
-          <Card className='border-teal-100'>
-            <CardHeader className='pb-2'>
-              <CardTitle className='text-sm font-medium text-teal-700 flex items-center gap-2'>
-                <BarChart3 className='h-4 w-4' />
-                Survey Results
-              </CardTitle>
-            </CardHeader>
-            <CardContent>
-              <div className='text-2xl font-bold'>
-                {Math.round(
-                  (highLevelCategoryStats.survey.completed / highLevelCategoryStats.survey.total) *
-                    100,
-                )}
-                %
-              </div>
-              <p className='text-xs text-muted-foreground'>
-                {highLevelCategoryStats.survey.completed} of {highLevelCategoryStats.survey.total}{' '}
-                checks complete
-              </p>
-            </CardContent>
-          </Card>
-        </div>
-      </CardContent>
-    </Card>
-  )
-}
 
 export const StudentDataCheck = (): JSX.Element => {
   const { phaseId } = useParams<{ phaseId: string }>()
